refactor: code-simplifier round 3
Four parallel simplifier passes plus a manual cleanup of providerConfig.

Net -474 LoC across 33 modified files; 3 new modules added.

Slices:

1. Tools/utils gravity-well split (-510 LoC in utils.ts; +520 across new
   modules). Split src/tools/utils.ts into pathResolution.ts (path-
   resolution helpers + assertWritable + resolveAndFormat),
   formatting.ts (formatFileView, pluralize, paginateToolListing,
   formatToolOutput, etc.), and attachments.ts (buildFileAttachment).
   utils.ts retains unrelated leftovers (createGlobMatcher,
   countOccurrences, requireField, wrapApiCall). 23 call sites updated
   mechanically. Behavior byte-identical; every compound guard preserved.

2. Latex small dedups. Private resolveFigureAbsolutePath helper for
   duplicate path construction in LatexMediaManager. getLocationPath
   inlined at 4 sites in TikzPictureManager. collectBibliographyPaths
   extracted into latexParsingUtils.ts; both callers retain their
   separate filesystem-API and existence-filter semantics.

3. Utils polish (partial). baseFS uses normalizeLineEndings helper at
   both inline sites. providerConfig: 7 of 10 simple getters consolidated
   via getStateValue factory. taskRunStorage legacy fallback skipped --
   recent (Feb 2026) consolidation; real installs may still hold legacy
   data.

4. utils/system/toolUtils.ts vscode boundary fix. Removed vscode import
   from a vscode-free zone. Injectable ToolMissingHandler set by
   extension.ts at activation, mirroring the existing setExtensionChecker
   pattern in externalToolDefs.ts. Zero caller changes (return shape and
   showError parameter unchanged).

5. providerConfig.ts deeper cleanup (manual). Replaced two parallel key
   maps (STREAMING_KEY, ENDPOINT_KEY) plus a hardcoded switch in
   getProviderKeyUrl with a single PROVIDERS registry entry per
   provider, including optional region overrides (displayName,
   keyUrlWhenSet, keyUrlWhenUnset). getProviderDisplayName and
   getProviderKeyUrl are now data-driven. Region getters
   (getDashScopeUseChina, getMiniMaxUseChina, getGLMUseChina) become
   1-line facades over the registry. Dropped setAnthropicDynamicFiltering
   (zero external callers). entry()/regionSet()/read() helpers
   eliminate per-getter boilerplate.

Verification: npm run typecheck clean. npm run lint reports the
same single pre-existing warning in sessionDescription.ts; confirmed
unchanged via stash.
