+++ src/extension.ts
@@ -72,6 +72,7 @@ import { setLeanVscodeServices } from '@tools/lean/leanVscodeServices';
 import { setOpenBuildDisplay } from '@tools/approval/latexPreview';
 import { StorageFS } from '@utils/files';
 import { getConfig } from '@utils/config';
+import { setToolMissingHandler } from '@utils/system';
 import { TASK_RUNS_DIR } from '@utils/files/taskRunStorage';
 
 // Local imports - components
@@ -240,6 +241,14 @@ export async function activate(context: vscode.ExtensionContext) {
   initializeNativeToolEditApproval(context);
   setLeanVscodeServices(leanVscodeIntegration);
   setExtensionChecker((id) => vscode.extensions.getExtension(id) !== undefined);
+  setToolMissingHandler(async (message, openDocsCommand) => {
+    const actions = openDocsCommand ? ['View Installation Guide'] : [];
+    const choice = await vscode.window.showErrorMessage(message, ...actions);
+    if (choice === 'View Installation Guide' && openDocsCommand) {
+      const [command, ...args] = openDocsCommand.split(',');
+      void vscode.commands.executeCommand(command, ...args);
+    }
+  });
   setSetupPlatform({
     secrets: {
       providers: SecretManager.API_PROVIDERS,
