+++ src/tools/arxiv/ArxivDownloadTool.ts
@@ -6,7 +6,7 @@ import { toErrorMessage } from '@common/errors';
 import { arxivProcessor } from '@latex/arxivProcessor';
 import { LsTool } from '@tools/ls';
 import { ToolError, type ToolResult } from '@tools/result';
-import { formatToolOutput } from '@tools/utils';
+import { formatToolOutput } from '@tools/formatting';
 import { defineTool } from '@tools/core/define';
 import { WorkspaceFS } from '@utils/files';
 import { toPosixPath } from '@utils/core/pathCore';
