+++ src/tools/arxiv/ArxivSearchTool.ts
@@ -10,7 +10,8 @@ import {
 import { z } from 'zod';
 
 // Local imports
-import { pluralize, requireNonEmptyString, wrapApiCall } from '@tools/utils';
+import { pluralize } from '@tools/formatting';
+import { requireNonEmptyString, wrapApiCall } from '@tools/utils';
 import { ARXIV_CONSTANTS } from '@tools/citation/constants';
 import { waitForRateLimit } from '@tools/citation/rateLimiter';
 import { defineTool } from '@tools/core/define';
