+++ src/tools/citation/CrossrefSearchTool.ts
@@ -9,7 +9,8 @@ import { z } from 'zod';
 
 // Local imports
 import { ToolError } from '@tools/result';
-import { pluralize, requireNonEmptyString, wrapApiCall } from '@tools/utils';
+import { pluralize } from '@tools/formatting';
+import { requireNonEmptyString, wrapApiCall } from '@tools/utils';
 import { defineTool } from '@tools/core/define';
 
 // Local file imports
