+++ src/tools/formatting.ts
@@ -1,0 +1,181 @@
+/** Maximum lines returned in a single file view before truncation. */
+export const READ_FILE_MAX_LINES = 2000;
+
+/** Default width for line number padding */
+const LINE_NUMBER_WIDTH = 6;
+
+/**
+ * Format lines with line numbers for display in tool output.
+ * @param lines - Array of lines to format
+ * @param startingLine - 1-based line number for the first line (default: 1)
+ * @param width - Padding width for line numbers (default: 6)
+ * @returns Array of formatted lines with line number prefix and tab separator
+ */
+export function formatLinesWithNumbers(
+  lines: string[],
+  startingLine: number = 1,
+  width: number = LINE_NUMBER_WIDTH,
+): string[] {
+  return lines.map((line, index) => {
+    const lineNumber = startingLine + index;
+    const prefix = lineNumber.toString().padStart(width, ' ');
+    return `${prefix}\t${line}`;
+  });
+}
+
+// ============================================================================
+// Shared file-view formatting
+// ============================================================================
+
+export interface FileViewOptions {
+  /** Display path used in the summary (e.g., "src/foo.ts" or "/memories/notes.md"). */
+  path: string;
+  /** All lines of the file. */
+  lines: string[];
+  /**
+   * Optional 1-based `[start, end]` (both inclusive) line range.
+   * When omitted or nullish, the full file is shown.
+   * Values are clamped to the file bounds automatically.
+   */
+  viewRange?: [number, number] | null;
+  /** Max visible lines before truncation. Defaults to READ_FILE_MAX_LINES (2000). */
+  maxLines?: number;
+  /** Optional suffix appended to the summary (e.g., memory metadata). */
+  summarySuffix?: string;
+}
+
+export interface FileViewResult {
+  [key: string]: unknown;
+  output: string;
+  summary: string;
+}
+
+/**
+ * Shared pipeline for displaying file content with line numbers, truncation,
+ * and a standardised "Read …" summary.  Used by read_file, text_editor view,
+ * memory view, and executions readFile.
+ */
+export function formatFileView({
+  path: filePath,
+  lines,
+  viewRange,
+  maxLines = READ_FILE_MAX_LINES,
+  summarySuffix = '',
+}: FileViewOptions): FileViewResult {
+  const totalLines = lines.length;
+  const rangeProvided = viewRange != null;
+  const startLine = Math.max(viewRange?.[0] ?? 1, 1);
+  const endLine = Math.min(viewRange?.[1] ?? totalLines, totalLines);
+  const rangeSize = Math.max(endLine - startLine + 1, 0);
+  const truncated = rangeSize > maxLines;
+  const sliceEnd = startLine - 1 + Math.min(rangeSize, maxLines);
+  const visibleLines = lines.slice(startLine - 1, sliceEnd);
+  const visibleCount = visibleLines.length;
+
+  // -- output ---------------------------------------------------------------
+  const segments: string[] = [];
+  if (visibleLines.length > 0) {
+    segments.push(formatLinesWithNumbers(visibleLines, startLine).join('\n'));
+  }
+  if (truncated) {
+    segments.push(`...(truncated, ${rangeSize - maxLines} more lines)`);
+  }
+
+  // -- summary --------------------------------------------------------------
+  let summary: string;
+
+  if (visibleCount === 0) {
+    const reason =
+      totalLines === 0 ? 'file is empty' : 'no lines in requested range';
+    summary = `Read ${filePath} (${reason})`;
+  } else {
+    const lastVisibleLine = startLine + visibleCount - 1;
+    const isFullRead =
+      !rangeProvided &&
+      !truncated &&
+      startLine === 1 &&
+      lastVisibleLine === totalLines;
+
+    if (isFullRead) {
+      summary = `Read ${filePath}`;
+    } else {
+      const rangeLabel =
+        startLine === lastVisibleLine
+          ? `line ${startLine}`
+          : `lines ${startLine}-${lastVisibleLine}`;
+      summary = `Read ${rangeLabel} of ${filePath}`;
+    }
+  }
+
+  if (summarySuffix) {
+    summary += summarySuffix;
+  }
+
+  return { output: segments.join('\n'), summary };
+}
+
+/**
+ * Pluralize a word based on count.
+ * Returns the singular form for count === 1, plural form otherwise.
+ */
+export function pluralize(
+  count: number,
+  singular: string,
+  plural?: string,
+): string {
+  return count === 1 ? singular : (plural ?? `${singular}s`);
+}
+
+/**
+ * Format a result count with proper pluralization.
+ * Example: formatResultCount(3, 'result') returns "3 results"
+ */
+export function formatResultCount(
+  count: number,
+  singular: string,
+  plural?: string,
+): string {
+  return `${count} ${pluralize(count, singular, plural)}`;
+}
+
+/**
+ * Format tool output with a header and content.
+ */
+export function formatToolOutput(
+  header: string,
+  content: string | string[] | null,
+  noMatchesText: string = '(no entries)',
+): string {
+  if (!content || (Array.isArray(content) && content.length === 0)) {
+    return `${header}\n${noMatchesText}`;
+  }
+  const body = Array.isArray(content) ? content.join('\n') : content;
+  return `${header}\n${body}`;
+}
+
+// ============================================================================
+// Offset-based pagination for tool listings
+// ============================================================================
+
+/**
+ * Slice an array by offset/limit and build display header + "next page" hint.
+ *
+ * All entries are loaded first then sliced client-side — this bounds the
+ * text returned to the model, not the I/O cost of building the listing.
+ */
+export function paginateToolListing<T>(
+  entries: readonly T[],
+  offset: number,
+  limit: number,
+): { page: readonly T[]; start: number; end: number; total: number } {
+  const total = entries.length;
+  const safeOffset = total > 0 ? Math.min(offset, total - 1) : 0;
+  const page = entries.slice(safeOffset, safeOffset + limit);
+  return { page, start: safeOffset + 1, end: safeOffset + page.length, total };
+}
+
+/** Format a "N more — use offset: X" hint, or empty string if no more pages. */
+export function formatPaginationHint(end: number, total: number): string {
+  if (end >= total) return '';
+  return `\n(${total - end} more — use offset: ${end} to see next page)`;
+}
