+++ src/tools/latex/ExtractTikzFiguresTool.ts
@@ -4,7 +4,7 @@ import { z } from 'zod';
 // Local imports - tools
 import { tikzPictureManager } from '@latex/TikzPictureManager';
 import { type ToolFileAttachment } from '@tools/result';
-import { formatResultCount, formatToolOutput } from '@tools/utils';
+import { formatResultCount, formatToolOutput } from '@tools/formatting';
 import { defineTool } from '@tools/core/define';
 import { pathToLocation } from '@utils/files';
 import {
