+++ src/tools/utils.ts
@@ -1,6 +1,3 @@
-// Standard library imports
-import * as path from 'path';
-
 // Third-party imports
 import { Minimatch } from 'minimatch';
 
@@ -8,221 +5,7 @@ import { Minimatch } from 'minimatch';
 import { toErrorMessage } from '@common/errors';
 
 // Local imports - tools
-import { isOversizedImage, MANY_IMAGE_MAX_DIMENSION } from '@tools/imageUtils';
-import { ToolError, type ToolFileAttachment } from '@tools/result';
-
-// Local imports - core utilities
-import { isNonEmptyString } from '@utils/core';
-import { WorkspaceFS, getMimeType } from '@utils/files';
-import { findExternalRoot } from '@utils/files/externalRoots';
-import { locatePathInRoot } from '@utils/files/workspaceRoot';
-import { toPosixPath } from '@utils/core/pathCore';
-
-export interface WorkspacePathResolution {
-  relative: string;
-  absolute: string;
-  /**
-   * The path to pass to filesystem operations.
-   * Absolute when operating outside the workspace (e.g. a worktree),
-   * workspace-relative otherwise (for WorkspaceFS compatibility).
-   */
-  fsPath: string;
-  /**
-   * When the resolved path falls inside a registered external root, this
-   * describes that root (label, writable flag). Undefined for workspace paths
-   * and for any external path outside the allowlist.
-   */
-  external?: { root: string; writable: boolean; label: string };
-}
-
-/** Trim and validate a working_directory value. Must be absolute if provided. */
-export function parseWorkingDirectory(
-  raw: string | null | undefined,
-): string | undefined {
-  const trimmed = raw?.trim() || undefined;
-  if (trimmed && !path.isAbsolute(trimmed)) {
-    throw new ToolError(
-      `working_directory must be an absolute path, got: ${trimmed}`,
-    );
-  }
-  return trimmed;
-}
-
-/**
- * Resolve a potentially absolute or relative path against a root directory.
- *
- * When `root` is provided, paths are resolved against that directory instead
- * of the workspace root. This supports operating in git worktrees or other
- * directories outside the main workspace.
- *
- * Thin policy wrapper around WorkspaceFS.locatePath() / locatePathInRoot()
- * that throws ToolError when the path escapes the root. Tools use this;
- * non-tool code should call WorkspaceFS.locatePath() directly.
- */
-export function resolveWorkspaceRelativePath(
-  targetPath?: string,
-  root?: string,
-): WorkspacePathResolution {
-  const trimmed = targetPath?.trim();
-  const input = !trimmed || trimmed === '.' ? '' : trimmed;
-
-  if (root) {
-    // Absolute paths need special handling — locatePathInRoot only works with relative paths.
-    if (input && path.isAbsolute(input)) {
-      // Normalize to POSIX separators so .relative is consistent across platforms
-      // (locatePathInRoot and WorkspaceFS.locatePath already do this).
-      const relative = path.relative(root, input).replaceAll('\\', '/');
-      if (relative.startsWith('..') || path.isAbsolute(relative)) {
-        const external = externalAllowance(input);
-        if (external) {
-          return external;
-        }
-        throw new ToolError('Path must stay within the working directory.');
-      }
-      return annotateExternalPermission({
-        relative: relative || '.',
-        absolute: input,
-        fsPath: input,
-      });
-    }
-    const resolved = locatePathInRoot(root, input);
-    if (resolved.kind === 'external') {
-      // `annotateExternal` in `locatePathInRoot` already consulted the
-      // registry, so reuse that match instead of paying for a second lookup.
-      if (resolved.allowed) {
-        return externalResolution(resolved.absolutePath, resolved.allowed);
-      }
-      throw new ToolError('Path must stay within the working directory.');
-    }
-    const relative = resolved.relativePath || '.';
-    return annotateExternalPermission({
-      relative,
-      absolute: resolved.absolutePath,
-      fsPath: resolved.absolutePath,
-    });
-  }
-
-  if (!WorkspaceFS.getPath()) {
-    // No workspace — fall back to the allowlist so agent-dir calls still work.
-    if (input && path.isAbsolute(input)) {
-      const external = externalAllowance(input);
-      if (external) {
-        return external;
-      }
-    }
-    throw new ToolError('Workspace path is not available.');
-  }
-
-  const resolved = WorkspaceFS.locatePath(input);
-
-  if (resolved.kind === 'external') {
-    if (resolved.allowed) {
-      return externalResolution(resolved.absolutePath, resolved.allowed);
-    }
-    throw new ToolError('Path must stay within the workspace.');
-  }
-
-  const relative = resolved.relativePath || '.';
-  return { relative, absolute: resolved.absolutePath, fsPath: relative };
-}
-
-/**
- * Build a WorkspacePathResolution for an absolute path that sits inside a
- * registered external root, using a pre-resolved allowlist match.
- *
- * `relative` is set to the full absolute path so the display (rendered via
- * `toPosixPath(relative)`) unambiguously signals an external operation —
- * agents and users should never confuse an external write with a workspace
- * write, even when file basenames collide.
- */
-function externalResolution(
-  absolutePath: string,
-  match: { absolutePath: string; writable: boolean; label: string },
-): WorkspacePathResolution {
-  return {
-    relative: absolutePath,
-    absolute: absolutePath,
-    fsPath: absolutePath,
-    external: {
-      root: match.absolutePath,
-      writable: match.writable,
-      label: match.label,
-    },
-  };
-}
-
-/**
- * Build a WorkspacePathResolution for an absolute path that sits inside a
- * registered external root. Returns undefined when no root matches.
- */
-function externalAllowance(
-  absolutePath: string,
-): WorkspacePathResolution | undefined {
-  const match = findExternalRoot(absolutePath);
-  if (!match) return undefined;
-  return externalResolution(absolutePath, match);
-}
-
-/**
- * Attach external-root permission metadata when a resolution (from the
- * working_directory branch) happens to land inside a registered root.
- *
- * Without this, a subagent launched with `working_directory` set to a
- * read-only external root (e.g. the built-in agents dir) could bypass
- * `assertWritable` by addressing files with paths relative to `root` —
- * the in-root branches return without touching the allowlist otherwise.
- *
- * Preserves `relative`/`absolute`/`fsPath` as the caller already built
- * them so display and I/O remain unchanged; only `external` is added.
- */
-function annotateExternalPermission(
-  resolution: WorkspacePathResolution,
-): WorkspacePathResolution {
-  if (resolution.external) return resolution;
-  const match = findExternalRoot(resolution.absolute);
-  if (!match) return resolution;
-  return {
-    ...resolution,
-    external: {
-      root: match.absolutePath,
-      writable: match.writable,
-      label: match.label,
-    },
-  };
-}
-
-/**
- * Throw a ToolError when the resolved path points into a read-only external
- * root. Workspace paths and writable externals pass through. Call this from
- * write/edit tools immediately before requesting approval.
- */
-export function assertWritable(
-  resolved: WorkspacePathResolution,
-  displayPath: string,
-): void {
-  if (resolved.external && !resolved.external.writable) {
-    throw new ToolError(
-      `Cannot write ${displayPath}: ${resolved.external.label} is read-only.`,
-    );
-  }
-}
-
-/**
- * Join a workspace-relative base path with a child segment and ensure the
- * result stays within the root directory.
- */
-export function joinWorkspaceRelativePath(
-  baseRelative: string,
-  child: string,
-  root?: string,
-): WorkspacePathResolution {
-  // Use posix.join since baseRelative and child are POSIX-normalized.
-  // path.join would reintroduce backslashes on Windows.
-  return resolveWorkspaceRelativePath(
-    path.posix.join(baseRelative || '.', child),
-    root,
-  );
-}
+import { ToolError } from '@tools/result';
 
 /**
  * Compile a glob pattern into a matcher that operates on POSIX-style paths.
@@ -251,177 +34,6 @@ export function countOccurrences(haystack: string, needle: string): number {
     index = haystack.indexOf(needle, index + needle.length);
   }
   return count;
-}
-
-/** Maximum lines returned in a single file view before truncation. */
-export const READ_FILE_MAX_LINES = 2000;
-
-/** Default width for line number padding */
-const LINE_NUMBER_WIDTH = 6;
-
-/**
- * Format lines with line numbers for display in tool output.
- * @param lines - Array of lines to format
- * @param startingLine - 1-based line number for the first line (default: 1)
- * @param width - Padding width for line numbers (default: 6)
- * @returns Array of formatted lines with line number prefix and tab separator
- */
-export function formatLinesWithNumbers(
-  lines: string[],
-  startingLine: number = 1,
-  width: number = LINE_NUMBER_WIDTH,
-): string[] {
-  return lines.map((line, index) => {
-    const lineNumber = startingLine + index;
-    const prefix = lineNumber.toString().padStart(width, ' ');
-    return `${prefix}\t${line}`;
-  });
-}
-
-// ============================================================================
-// Shared file-view formatting
-// ============================================================================
-
-export interface FileViewOptions {
-  /** Display path used in the summary (e.g., "src/foo.ts" or "/memories/notes.md"). */
-  path: string;
-  /** All lines of the file. */
-  lines: string[];
-  /**
-   * Optional 1-based `[start, end]` (both inclusive) line range.
-   * When omitted or nullish, the full file is shown.
-   * Values are clamped to the file bounds automatically.
-   */
-  viewRange?: [number, number] | null;
-  /** Max visible lines before truncation. Defaults to READ_FILE_MAX_LINES (2000). */
-  maxLines?: number;
-  /** Optional suffix appended to the summary (e.g., memory metadata). */
-  summarySuffix?: string;
-}
-
-export interface FileViewResult {
-  [key: string]: unknown;
-  output: string;
-  summary: string;
-}
-
-/**
- * Shared pipeline for displaying file content with line numbers, truncation,
- * and a standardised "Read …" summary.  Used by read_file, text_editor view,
- * memory view, and executions readFile.
- */
-export function formatFileView({
-  path: filePath,
-  lines,
-  viewRange,
-  maxLines = READ_FILE_MAX_LINES,
-  summarySuffix = '',
-}: FileViewOptions): FileViewResult {
-  const totalLines = lines.length;
-  const rangeProvided = viewRange != null;
-  const startLine = Math.max(viewRange?.[0] ?? 1, 1);
-  const endLine = Math.min(viewRange?.[1] ?? totalLines, totalLines);
-  const rangeSize = Math.max(endLine - startLine + 1, 0);
-  const truncated = rangeSize > maxLines;
-  const sliceEnd = startLine - 1 + Math.min(rangeSize, maxLines);
-  const visibleLines = lines.slice(startLine - 1, sliceEnd);
-  const visibleCount = visibleLines.length;
-
-  // -- output ---------------------------------------------------------------
-  const segments: string[] = [];
-  if (visibleLines.length > 0) {
-    segments.push(formatLinesWithNumbers(visibleLines, startLine).join('\n'));
-  }
-  if (truncated) {
-    segments.push(`...(truncated, ${rangeSize - maxLines} more lines)`);
-  }
-
-  // -- summary --------------------------------------------------------------
-  let summary: string;
-
-  if (visibleCount === 0) {
-    const reason =
-      totalLines === 0 ? 'file is empty' : 'no lines in requested range';
-    summary = `Read ${filePath} (${reason})`;
-  } else {
-    const lastVisibleLine = startLine + visibleCount - 1;
-    const isFullRead =
-      !rangeProvided &&
-      !truncated &&
-      startLine === 1 &&
-      lastVisibleLine === totalLines;
-
-    if (isFullRead) {
-      summary = `Read ${filePath}`;
-    } else {
-      const rangeLabel =
-        startLine === lastVisibleLine
-          ? `line ${startLine}`
-          : `lines ${startLine}-${lastVisibleLine}`;
-      summary = `Read ${rangeLabel} of ${filePath}`;
-    }
-  }
-
-  if (summarySuffix) {
-    summary += summarySuffix;
-  }
-
-  return { output: segments.join('\n'), summary };
-}
-
-/**
- * Pluralize a word based on count.
- * Returns the singular form for count === 1, plural form otherwise.
- */
-export function pluralize(
-  count: number,
-  singular: string,
-  plural?: string,
-): string {
-  return count === 1 ? singular : (plural ?? `${singular}s`);
-}
-
-/**
- * Format a result count with proper pluralization.
- * Example: formatResultCount(3, 'result') returns "3 results"
- */
-export function formatResultCount(
-  count: number,
-  singular: string,
-  plural?: string,
-): string {
-  return `${count} ${pluralize(count, singular, plural)}`;
-}
-
-/**
- * Format tool output with a header and content.
- */
-export function formatToolOutput(
-  header: string,
-  content: string | string[] | null,
-  noMatchesText: string = '(no entries)',
-): string {
-  if (!content || (Array.isArray(content) && content.length === 0)) {
-    return `${header}\n${noMatchesText}`;
-  }
-  const body = Array.isArray(content) ? content.join('\n') : content;
-  return `${header}\n${body}`;
-}
-
-/**
- * Common pattern for resolving and formatting paths.
- * Returns `path` (resolution with relative/absolute) and `display` (formatted string).
- */
-export function resolveAndFormat(
-  targetPath?: string,
-  root?: string,
-): {
-  path: WorkspacePathResolution;
-  display: string;
-} {
-  const path = resolveWorkspaceRelativePath(targetPath, root);
-  const display = toPosixPath(path.relative);
-  return { path, display };
 }
 
 /**
@@ -469,124 +81,4 @@ export async function wrapApiCall<T>(
   } catch (error) {
     throw new ToolError(`${errorPrefix}: ${toErrorMessage(error)}`);
   }
-}
-
-// ============================================================================
-// Offset-based pagination for tool listings
-// ============================================================================
-
-/**
- * Slice an array by offset/limit and build display header + "next page" hint.
- *
- * All entries are loaded first then sliced client-side — this bounds the
- * text returned to the model, not the I/O cost of building the listing.
- */
-export function paginateToolListing<T>(
-  entries: readonly T[],
-  offset: number,
-  limit: number,
-): { page: readonly T[]; start: number; end: number; total: number } {
-  const total = entries.length;
-  const safeOffset = total > 0 ? Math.min(offset, total - 1) : 0;
-  const page = entries.slice(safeOffset, safeOffset + limit);
-  return { page, start: safeOffset + 1, end: safeOffset + page.length, total };
-}
-
-/** Format a "N more — use offset: X" hint, or empty string if no more pages. */
-export function formatPaginationHint(end: number, total: number): string {
-  if (end >= total) return '';
-  return `\n(${total - end} more \u2014 use offset: ${end} to see next page)`;
-}
-
-export interface BuildFileAttachmentOptions {
-  /** Path to a workspace file (relative or absolute) */
-  filePath: string;
-  /** Optional description surfaced to the model */
-  description?: string;
-  /** Override detected MIME type */
-  mimeType?: string;
-  /** Include base64 data in the attachment */
-  includeBase64?: boolean;
-  /** Maximum allowed file size in bytes */
-  maxBytes?: number;
-  /** Optional root directory override (e.g. a git worktree) */
-  root?: string;
-  /**
-   * Pre-resolved path. When provided, skips the internal resolveAndFormat()
-   * call — use this to avoid double-resolution when the caller already
-   * resolved the path (e.g. ReadTool).
-   */
-  resolved?: WorkspacePathResolution;
-}
-
-const DEFAULT_ATTACHMENT_MAX_BYTES = 15 * 1024 * 1024; // 15 MiB
-
-/**
- * Build a tool attachment by reading a workspace file and packaging metadata.
- */
-export async function buildFileAttachment({
-  filePath,
-  description,
-  mimeType,
-  includeBase64 = false,
-  maxBytes = DEFAULT_ATTACHMENT_MAX_BYTES,
-  root,
-  resolved,
-}: BuildFileAttachmentOptions): Promise<ToolFileAttachment> {
-  if (!isNonEmptyString(filePath)) {
-    throw new ToolError('Attachment path must be provided.');
-  }
-
-  const { path, display } = resolved
-    ? { path: resolved, display: toPosixPath(resolved.relative) }
-    : resolveAndFormat(filePath, root);
-  const exists = await WorkspaceFS.exists(path.fsPath);
-  if (!exists) {
-    throw new ToolError(`Attachment not found: ${display}`);
-  }
-
-  const stats = await wrapApiCall(
-    () => WorkspaceFS.stat(path.fsPath),
-    `Failed to inspect attachment ${display}`,
-  );
-
-  if (stats.size > maxBytes) {
-    const limitMb = (maxBytes / (1024 * 1024)).toFixed(1);
-    throw new ToolError(
-      `Attachment ${display} exceeds maximum size of ${limitMb} MiB.`,
-    );
-  }
-
-  const buffer = await wrapApiCall(
-    () => WorkspaceFS.readBytes(path.fsPath),
-    `Failed to read attachment ${display}`,
-  );
-
-  const inferredMime =
-    mimeType ?? getMimeType(path.fsPath) ?? 'application/octet-stream';
-
-  // Strip binary data from oversized images to prevent non-retryable API 400 errors.
-  // Downstream handlers see no bytes → metadata-only fallback with read_file hint.
-  if (inferredMime.startsWith('image/') && isOversizedImage(buffer)) {
-    buffer.fill(0);
-    return {
-      path: display,
-      mimeType: inferredMime,
-      description:
-        (description ? `${description} — ` : '') +
-        `Image exceeds ${MANY_IMAGE_MAX_DIMENSION}px dimension limit; binary data stripped`,
-    };
-  }
-
-  const base64Data = includeBase64 ? buffer.toString('base64') : undefined;
-  const bytes = Uint8Array.from(buffer);
-  buffer.fill(0);
-
-  return {
-    path: display,
-    mimeType: inferredMime,
-    bytes,
-    ...(description && { description }),
-    ...(base64Data && { base64Data }),
-  };
 }
