+++ src/tools/zotero/ZoteroExportTool.ts
@@ -10,7 +10,7 @@ import { z } from 'zod';
 
 // Local imports - core
 import { ToolError } from '@tools/result';
-import { pluralize } from '@tools/utils';
+import { pluralize } from '@tools/formatting';
 import { defineTool } from '@tools/core/define';
 
 // Local imports - zotero
